Google maps. API KEY
Google maps need a key for create it custom #5

- Bug fixed. Map not load.
For the map to load it is necessary to use an api that from google maps.
Go here and read the instructions:
https://developers.google.com/maps/documentation/geocoding/get-api-key
Change in 'UserMap.vue' the string 'XXXXXX' for the your API_KEY

- Reduce zoom to have a wider view of the user's location.

- Bug fixed. CodeSandBox has problems with empty style tags. You add something on the label or remove it.
Removed the empty styles of the 'User.vue'.