BUGFIX. Annoying noise when adding/removing extra spaces
Error:
- I was really frustrated.
When adding or removing extra spaces in Visual Studio Code (VSC), an annoying noise was being triggered.

Solution:
- Disable the VSC setting: `editor.accessibilitySupport`:
This setting enables certain accessibility-related features, such as notification sounds when there are errors in the code.
Disabling this option may be useful for users who do not need or prefer these accessibility features and want to avoid unnecessary sounds and distractions.

References:
- How to turn off (or on) sounds from Visual Studio Code?
https://stackoverflow.com/a/71042502
- Current known issues
https://code.visualstudio.com/docs/editor/accessibility#_current-known-issues