fix(MapGoogle): migrate code to use `@googlemaps/js-api-loader` package
Errors:
- InvalidValueError: Unknown property 'marker' of AdvancedMarkerElement.
- InvalidValueError: Unknown property 'icon' of AdvancedMarkerElement.
- Error: <svg> attribute viewBox: Expected number, "undefined".
- Error: <svg> attribute width: Expected length, "undefined".
- Error: <svg> attribute height: Expected length, "undefined".
- <gmp-advanced-marker>: Please use addEventListener('gmp-click', ...) instead of addEventListener('click', ...).
- The map was initialized without a valid Map ID, so advanced markers were disabled.
- Google Maps JavaScript API: A Map's styles property cannot be set when a mapId is present. When a mapId is present, map styles are controlled via the cloud console. Please see documentation at https://developers.google.com/maps/documentation/javascript/styling#cloud_tooling

Solution:
- In `MapGoogleLoader` component:
Replace import `google-maps-api-loader` with `@googlemaps/js-api-loader`, which provides typed loading and native support for ES modules.
Replace `googleMapsApiLoader()` with `Loader` class and use `importLibrary()` to load `maps` and `marker` libraries.
Specify `"version": "beta"` property in the configuration to unlock access to `AdvancedMarkerElement()` and `importLibrary()`, which are not available in the stable version. This avoids errors caused by initializing the loader more than once with different options.
- In `MapGoogleMarker` component:
Replace the legacy `Marker()` constructor with `AdvancedMarkerElement()` and remove deprecated properties such as `marker` and `icon`.
Use the `content` property instead of `icon` because `AdvancedMarkerElement()` requires a DOM node rather than a plain object. To solve this, create a new `createSvgIcon()` method to construct the `<svg>` tag and add `viewBox`, `width`, and `height` properties into `MAP_GOOGLE_MARKER_ICON` constant, which are now mandatory but were previously implicit in the marker config.
Set `"gmpClickable": true` property and update the event listener from click to `gmp-click` according to the new API requirements.
- In `MapGoogle` component, `.env.local` and `.env.local.sample` files:
Add a `mapId` property to the `mapConfig()` function, using the environment variable `VUE_APP_ID_GOOGLE_MAPS` as its value. This is required to enable support for advanced features such as `AdvancedMarkerElement()`.
Since the `mapId` is a sensitive key, it must be added to the `.env.local` file for local use, included in `.env.local.sample` to assist future developers, and declared as a `GitHub Actions secret` to enable production `deployment` workflows.
Since custom styles cannot be set when a `mapId` is present, remove the `styles` property (`MAP_GOOGLE_CONFIG_STYLES`). From now on, map styling must be configured through the `Google Cloud Console`.

Reason:
- As of February 21st, 2024, google.maps.Marker is deprecated. Please use google.maps.marker.AdvancedMarkerElement instead. At this time, google.maps.Marker is not scheduled to be discontinued, but google.maps.marker.AdvancedMarkerElement is recommended over google.maps.Marker. While google.maps.Marker will continue to receive bug fixes for any major regressions, existing bugs in google.maps.Marker will not be addressed. At least 12 months notice will be given before support is discontinued. Please see https://developers.google.com/maps/deprecations for additional details and https://developers.google.com/maps/documentation/javascript/advanced-markers/migration for the migration guide.

References:
- https://www.npmjs.com/package/@googlemaps/js-api-loader
- https://developers.google.com/maps/documentation/javascript/advanced-markers/migration?hl=es-419
- https://developers.google.com/maps/documentation/javascript/reference/advanced-markers?hl=es-419#AdvancedMarkerElementOptions
- https://developers.google.com/maps/documentation/javascript/load-maps-js-api
- https://developers.google.com/maps/documentation/javascript/versions
- https://docs.github.com/en/actions/security-guides/encrypted-secrets
- https://developers.google.com/maps/documentation/javascript/styling#cloud_tooling
- https://console.cloud.google.com/google/maps-apis/studio/styles
- https://mapstyle.withgoogle.com/
- https://snazzymaps.com/